feat: Continue writing welcome file
