+++ battleship/spec/welcome_spec.rb
